completed

+++ src/modals/TaskModal.jsx
@@ -1,16 +1,18 @@
 import React, { useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import ellipsis from "../assets/icon-vertical-ellipsis.svg";
 import EllipsisMenu from "../components/EllipsisMenu";
+import ellipsis from "../assets/icon-vertical-ellipsis.svg";
 import Subtask from "../components/Subtask";
-import { deleteTask, setTaskStatus } from "../redux/boardsSlice";
-import DeleteModal from "./DeleteModal";
 import AddEditTaskModal from "./AddEditTaskModal";
+import DeleteModal from "./DeleteModal";
+import { deleteTask, setTaskStatus } from "../redux/boardsSlice";
 
-const TaskModal = ({ colIndex, taskIndex, setIsTaskModalOpen }) => {
+function TaskModal({ taskIndex, colIndex, setIsTaskModalOpen }) {
   const dispatch = useDispatch();
+  const [isEllipsisMenuOpen, setIsEllipsisMenuOpen] = useState(false);
+  const [isDeleteModalOpen, setIsDeleteModalOpen] = useState(false);
   const boards = useSelector((state) => state.boards);
-  const board = boards.find((board) => board.isActive);
+  const board = boards.find((board) => board.isActive === true);
   const columns = board.columns;
   const col = columns.find((col, i) => i === colIndex);
   const task = col.tasks.find((task, i) => i === taskIndex);
@@ -22,20 +24,12 @@ const TaskModal = ({ colIndex, taskIndex, setIsTaskModalOpen }) => {
       completed++;
     }
   });
+
   const [status, setStatus] = useState(task.status);
   const [newColIndex, setNewColIndex] = useState(columns.indexOf(col));
-  const [ellipsisMenuOpen, setEllipsisMenuOpen] = useState(false);
-  const [isDeleteModalOpen, setIsDeleteModalOpen] = useState(false);
-  const [isAddTaskModalOpen, setIsAddTaskModalOpen] = useState(false);
-
-  const setOpenEditModal = () => {
-    setIsAddTaskModalOpen(true);
-    setEllipsisMenuOpen(false);
-  };
-
-  const setOpenDeleteModal = () => {
-    setEllipsisMenuOpen(false);
-    setIsDeleteModalOpen(true);
+  const onChange = (e) => {
+    setStatus(e.target.value);
+    setNewColIndex(e.target.selectedIndex);
   };
 
   const onClose = (e) => {
@@ -53,40 +47,48 @@ const TaskModal = ({ colIndex, taskIndex, setIsTaskModalOpen }) => {
     setIsTaskModalOpen(false);
   };
 
-  const onChange = (e) => {
-    setStatus(e.target.value);
-    setNewColIndex(e.target.selectedIndex);
+  const onDeleteBtnClick = (e) => {
+    if (e.target.textContent === "Delete") {
+      dispatch(deleteTask({ taskIndex, colIndex }));
+      setIsTaskModalOpen(false);
+      setIsDeleteModalOpen(false);
+    } else {
+      setIsDeleteModalOpen(false);
+    }
   };
 
-  const onDeleteBtnClick = () => {
-    dispatch(deleteTask({ taskIndex, colIndex }));
-    setIsTaskModalOpen(false);
-    setIsDeleteModalOpen(false);
+  const [isAddTaskModalOpen, setIsAddTaskModalOpen] = useState(false);
+
+  const setOpenEditModal = () => {
+    setIsAddTaskModalOpen(true);
+    setIsEllipsisMenuOpen(false);
+  };
+
+  const setOpenDeleteModal = () => {
+    setIsEllipsisMenuOpen(false);
+    setIsDeleteModalOpen(true);
   };
 
   return (
     <div
       onClick={onClose}
-      className="fixed right-0 left-0 top-0 px-2 py-4 overflow-scroll
-       scrollbar-hide z-50 bottom-0 justify-center items-center flex bg-[#00000080] "
+      className=" fixed right-0 top-0 px-2 py-4 overflow-scroll scrollbar-hide  z-50 left-0 bottom-0 justify-center items-center flex dropdown "
     >
-      {/* Modal Section */}
-      <div
-        className="scrollbar-hide overflow-y-scroll max-h-[95vh] my-auto
-      bg-white dark:bg-[#2b2c37 text-black dark:text-white font-bold
-        shadow-md shadow-[#364e7e1a] max-w-md mx-auto w-full px-8 py-8 rounded-xl"
-      >
-        <div className="relative flex justify-between w-full items-center">
-          <h1 className="text-lg">{task.title}</h1>
+      {/* MODAL SECTION */}
+
+      <div className=" scrollbar-hide overflow-y-scroll max-h-[95vh]  my-auto  bg-white dark:bg-[#2b2c37] text-black dark:text-white font-bold shadow-md shadow-[#364e7e1a] max-w-md mx-auto  w-full px-8  py-8 rounded-xl">
+        <div className=" relative flex   justify-between w-full items-center">
+          <h1 className=" text-lg">{task.title}</h1>
+
           <img
             onClick={() => {
-              setEllipsisMenuOpen((state) => !state);
+              setIsEllipsisMenuOpen((prevState) => !prevState);
             }}
             src={ellipsis}
-            alt="Ellipsis"
-            className="cursor-pointer h-6"
+            alt="ellipsis"
+            className=" cursor-pointer h-6"
           />
-          {ellipsisMenuOpen && (
+          {isEllipsisMenuOpen && (
             <EllipsisMenu
               setOpenEditModal={setOpenEditModal}
               setOpenDeleteModal={setOpenDeleteModal}
@@ -94,42 +96,43 @@ const TaskModal = ({ colIndex, taskIndex, setIsTaskModalOpen }) => {
             />
           )}
         </div>
-
-        <p className="text-gray-500 font-semibold tracking-wide text-sm pt-6">
+        <p className=" text-gray-500 font-[600] tracking-wide text-xs pt-6">
           {task.description}
         </p>
 
-        <p className="pt-6 text-gray-500 tracking-widest text-sm">
-          Subtasks ({completed}) of {subtasks.length}
+        <p className=" pt-6 text-gray-500 tracking-widest text-sm">
+          Subtasks ({completed} of {subtasks.length})
         </p>
 
-        {/* Subtasks Section */}
-        <div className="mt-3 space-y-2">
-          {subtasks.map((subtask, index) => (
-            <Subtask
-              key={index}
-              index={index}
-              taskIndex={taskIndex}
-              colIndex={colIndex}
-            />
-          ))}
+        {/* subtasks section */}
+
+        <div className=" mt-3 space-y-2">
+          {subtasks.map((subtask, index) => {
+            return (
+              <Subtask
+                index={index}
+                taskIndex={taskIndex}
+                colIndex={colIndex}
+                key={index}
+              />
+            );
+          })}
         </div>
 
         {/* Current Status Section */}
 
         <div className="mt-8 flex flex-col space-y-3">
-          <label className="text-sm dark:text-white text-gray-500">
+          <label className="  text-sm dark:text-white text-gray-500">
             Current Status
           </label>
           <select
+            className=" select-status flex-grow px-4 py-2 rounded-md text-sm bg-transparent focus:border-0  border-[1px] border-gray-300 focus:outline-[#635fc7] outline-none"
             value={status}
             onChange={onChange}
-            className="select-status flex-grow px-4 py-2 rounded-md
-            bg-transparent focus:border-0 border border-gray-300 focus:outline-[#635fc7] outline-none"
           >
-            {columns.map((column, index) => (
-              <option key={index} className="status-option">
-                {column.name}
+            {columns.map((col, index) => (
+              <option className="status-options" key={index}>
+                {col.name}
               </option>
             ))}
           </select>
@@ -137,24 +140,23 @@ const TaskModal = ({ colIndex, taskIndex, setIsTaskModalOpen }) => {
       </div>
       {isDeleteModalOpen && (
         <DeleteModal
-          setIsDeleteModalOpen={setIsDeleteModalOpen}
           onDeleteBtnClick={onDeleteBtnClick}
-          title={task.title}
           type="task"
+          title={task.title}
         />
       )}
 
       {isAddTaskModalOpen && (
         <AddEditTaskModal
-          setOpenAddEditTask={setIsAddTaskModalOpen}
+          setIsAddTaskModalOpen={setIsAddTaskModalOpen}
+          setIsTaskModalOpen={setIsTaskModalOpen}
           type="edit"
           taskIndex={taskIndex}
           prevColIndex={colIndex}
-          setIsTaskModalOpen={setIsTaskModalOpen}
         />
       )}
     </div>
   );
-};
+}
 
 export default TaskModal;
